refactor(runtime): split persistence modules for size gate
